# docs/swagger.yaml
definitions:
  database.People:
    properties:
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      team:
        type: integer
    type: object
  database.User:
    properties:
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  main.PersonInput:
    properties:
      name:
        type: string
      role:
        type: string
    required:
    - name
    - role
    type: object
  main.RandomizeRequest:
    properties:
      options:
        $ref: '#/definitions/main.RandomizeRequestOpts'
      people:
        items:
          $ref: '#/definitions/main.PersonInput'
        minItems: 1
        type: array
      team_count:
        minimum: 1
        type: integer
    required:
    - people
    - team_count
    type: object
  main.RandomizeRequestOpts:
    type: object
  main.RandomizeResponse:
    properties:
      teams:
        items:
          $ref: '#/definitions/main.TeamGroup'
        type: array
      total:
        type: integer
    type: object
  main.TeamGroup:
    properties:
      members:
        items:
          $ref: '#/definitions/database.People'
        type: array
      team:
        type: integer
    type: object
  main.errorResponse:
    properties:
      error:
        type: string
    type: object
  main.googleAuthRequest:
    properties:
      client_id:
        type: string
      client_secret:
        type: string
      code:
        type: string
      redirect_url:
        type: string
      state:
        type: string
    required:
    - client_id
    - client_secret
    - code
    - redirect_url
    - state
    type: object
  main.loginRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    type: object
  main.loginResponse:
    properties:
      token:
        type: string
      user_id:
        type: integer
    type: object
  main.registerRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    type: object
  main.registerResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/database.User'
    type: object
info:
  contact: {}
paths:
  /v1/auth/google:
    post:
      consumes:
      - application/json
      description: Exchanges Google OAuth2 code for user info, upserts the user, and
        returns JWT
      parameters:
      - description: OAuth exchange request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.googleAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Google OAuth login/signup
      tags:
      - auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Issues JWT after verifying credentials
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Login with email/password
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, name, and password
      parameters:
      - description: Register Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.registerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Register a new user
      tags:
      - auth
  /v1/people/history:
    get:
      description: Returns saved randomizations for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RandomizeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Get saved team history
      tags:
      - people
  /v1/people/randomize:
    post:
      consumes:
      - application/json
      description: Shuffles people, assigns teams, optionally saves for authenticated
        users
      parameters:
      - description: Randomize request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.RandomizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RandomizeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Randomly assign people into teams
      tags:
      - people
  /v1/people/randomize/custom:
    post:
      consumes:
      - application/json
      description: Custom randomization of people into teams
      parameters:
      - description: Randomize request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.RandomizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RandomizeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Custom randomize (TODO)
      tags:
      - people
swagger: "2.0"